Changes to the chat files.
I have changed the files messages and chat. Pretty much changed the meaning of them but kept the same name on the files for now, a bit confusing. So that the chat file takes care of the messages in a file, and the message files takes care of the overivew of the chats. 
Also this is currently in progress. Not done anywhere and it's a mess. I wouldnt recomend working on these files since it might be messed up if we merge em in it's current state.